+++ backend_users/dev/tests/test_db_model/test_user_project_db_model.py
@@ -1,4 +1,5 @@
-from prod.db_models.user_db_model import UserDBModel, UserProjectDBModel
+from prod.db_models.user_db_model import UserDBModel
+from prod.db_models.user_project_db_model import UserProjectDBModel
 from dev.aux_test import recreate_db
